requirements.txt
Django==4.0.6
djangorestframework==3.13.1
django-environ==0.9.0
django-cors-headers==3.13.0
django-ckeditor==6.4.2
gunicorn==20.1.0
Pillow==9.2.0
whitenoise==6.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0